requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-PyMongo==2.3.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
PyJWT==1.7.1
pymongo==3.11.2
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.1
twilio==6.50.1
urllib3==1.26.2
Werkzeug==1.0.1
